Ensuring that the write-contents-hook returns nil
The auto-indent-whole-buffer was being called in a
write-contents-hook. It is important for a write-contents-hook to return
nil since according to the documentation if any function in the hook
returns a non-nil the file is considered already written and the rest of
the functions are not called and neither are the functions in
write-file-functions.